+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/transformers/InjectConstructorTransformer.kt
@@ -11,7 +11,6 @@ import dev.zacsweers.metro.compiler.ir.assignConstructorParamsToFields
 import dev.zacsweers.metro.compiler.ir.createIrBuilder
 import dev.zacsweers.metro.compiler.ir.dispatchReceiverFor
 import dev.zacsweers.metro.compiler.ir.finalizeFakeOverride
-import dev.zacsweers.metro.compiler.ir.implementsProviderType
 import dev.zacsweers.metro.compiler.ir.irExprBodySafe
 import dev.zacsweers.metro.compiler.ir.irInvoke
 import dev.zacsweers.metro.compiler.ir.irTemporary
@@ -27,7 +26,6 @@ import dev.zacsweers.metro.compiler.ir.typeAsProviderArgument
 import kotlin.collections.component1
 import kotlin.collections.component2
 import org.jetbrains.kotlin.descriptors.ClassKind
-import org.jetbrains.kotlin.ir.builders.IrBuilderWithScope
 import org.jetbrains.kotlin.ir.builders.irBlockBody
 import org.jetbrains.kotlin.ir.builders.irCallConstructor
 import org.jetbrains.kotlin.ir.builders.irGet
@@ -39,26 +37,20 @@ import org.jetbrains.kotlin.ir.declarations.IrConstructor
 import org.jetbrains.kotlin.ir.declarations.IrField
 import org.jetbrains.kotlin.ir.declarations.IrSimpleFunction
 import org.jetbrains.kotlin.ir.declarations.IrValueParameter
-import org.jetbrains.kotlin.ir.expressions.IrExpression
 import org.jetbrains.kotlin.ir.symbols.IrConstructorSymbol
-import org.jetbrains.kotlin.ir.symbols.IrFunctionSymbol
-import org.jetbrains.kotlin.ir.types.typeWith
 import org.jetbrains.kotlin.ir.util.callableId
 import org.jetbrains.kotlin.ir.util.classIdOrFail
 import org.jetbrains.kotlin.ir.util.companionObject
-import org.jetbrains.kotlin.ir.util.defaultType
 import org.jetbrains.kotlin.ir.util.file
 import org.jetbrains.kotlin.ir.util.functions
 import org.jetbrains.kotlin.ir.util.getAnnotation
 import org.jetbrains.kotlin.ir.util.getAnnotationStringValue
-import org.jetbrains.kotlin.ir.util.isFromJava
 import org.jetbrains.kotlin.ir.util.isObject
 import org.jetbrains.kotlin.ir.util.kotlinFqName
 import org.jetbrains.kotlin.ir.util.nestedClasses
 import org.jetbrains.kotlin.ir.util.packageFqName
 import org.jetbrains.kotlin.ir.util.parentAsClass
 import org.jetbrains.kotlin.ir.util.primaryConstructor
-import org.jetbrains.kotlin.ir.util.simpleFunctions
 import org.jetbrains.kotlin.name.CallableId
 import org.jetbrains.kotlin.name.ClassId
 
@@ -444,86 +436,5 @@ internal class InjectConstructorTransformer(
         }
       }
     return newInstanceFunction
-  }
-
-  sealed interface ConstructorInjectedFactory {
-    val factoryClass: IrClass
-    val invokeFunctionSymbol: IrFunctionSymbol
-
-    fun IrBuilderWithScope.invokeCreateExpression(
-      computeArgs: IrBuilderWithScope.(createFunction: IrSimpleFunction) -> List<IrExpression?>
-    ): IrExpression
-
-    class MetroFactory(override val factoryClass: IrClass) : ConstructorInjectedFactory {
-      override val invokeFunctionSymbol: IrFunctionSymbol
-        get() = factoryClass.requireSimpleFunction(Symbols.StringNames.INVOKE)
-
-      override fun IrBuilderWithScope.invokeCreateExpression(
-        computeArgs: IrBuilderWithScope.(IrSimpleFunction) -> List<IrExpression?>
-      ): IrExpression {
-        // Invoke its factory's create() function
-        val creatorClass =
-          if (factoryClass.isObject) {
-            factoryClass
-          } else {
-            factoryClass.companionObject()!!
-          }
-        val createFunction = creatorClass.requireSimpleFunction(Symbols.StringNames.CREATE)
-        val args = computeArgs(createFunction.owner)
-        return irInvoke(
-          dispatchReceiver = irGetObject(creatorClass.symbol),
-          callee = createFunction,
-          args = args,
-          typeHint = factoryClass.typeWith(),
-        )
-      }
-    }
-
-    class DaggerFactory(
-      private val metroContext: IrMetroContext,
-      override val factoryClass: IrClass,
-    ) : ConstructorInjectedFactory {
-      override val invokeFunctionSymbol: IrFunctionSymbol
-        get() = factoryClass.requireSimpleFunction(Symbols.StringNames.GET)
-
-      override fun IrBuilderWithScope.invokeCreateExpression(
-        computeArgs: IrBuilderWithScope.(createFunction: IrSimpleFunction) -> List<IrExpression?>
-      ): IrExpression {
-        // Anvil may generate the factory
-        val isJava = factoryClass.isFromJava()
-        val creatorClass =
-          if (isJava || factoryClass.isObject) {
-            factoryClass
-          } else {
-            factoryClass.companionObject()!!
-          }
-        val createFunction =
-          creatorClass
-            .simpleFunctions()
-            .first {
-              it.name == Symbols.Names.create || it.name == Symbols.Names.createFactoryProvider
-            }
-            .symbol
-        val args = computeArgs(createFunction.owner)
-        val createExpression =
-          irInvoke(
-            dispatchReceiver = if (isJava) null else irGetObject(creatorClass.symbol),
-            callee = createFunction,
-            args = args,
-            typeHint = factoryClass.typeWith(),
-          )
-
-        // Wrap in a metro provider if this is a provider
-        return if (factoryClass.defaultType.implementsProviderType(metroContext)) {
-          irInvoke(
-              extensionReceiver = createExpression,
-              callee = metroContext.symbols.daggerSymbols.asMetroProvider,
-            )
-            .apply { putTypeArgument(0, factoryClass.typeWith()) }
-        } else {
-          createExpression
-        }
-      }
-    }
   }
 }
